Interactive API: move ManagedState to its own file/module
[#173082951]
